requirements.txt
amqp==5.0.6
asgiref==3.2.10
billiard==3.6.4.0
celery==5.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1
django-celery-beat==2.2.0
django-celery-results==2.1.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-neomodel==0.0.6
django-restframework==0.0.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
importlib-metadata==4.6.1
kombu==5.1.0
Markdown==3.3.4
mysqlclient==2.0.3
neo4j-driver==4.1.1
neobolt==1.7.17
neomodel==4.0.3
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.1
redis==3.5.3
Shapely==1.7.1
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
git+git://github.com/shephexd/linchfin@develop